server/requirements.txt
youtube-transcript-api
google-generativeai
python-dotenv
pathlib
flask
flask-smorest
flask_cors
uuid
marshmallow
flask-sqlalchemy
sqlalchemy
flask_jwt_extended
passlib
gunicorn
pytube